Backend/backend-santander/requirements.txt
# Atlas al Crimen - Santander Backend
# Dependencias Python

# Web Framework
fastapi==0.123.5
uvicorn[standard]==0.38.0

# Base de datos
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
geoalchemy2==0.18.1

# Configuración
pyyaml==6.0.3
python-dotenv==1.2.1
pydantic-settings==2.12.0

# Utilidades
shapely==2.1.2
google-generativeai
